[NFC] Move dwarf::CFIProgram from DWARFDebugFrame.[h,cpp] to its own files
This is the first in a series of refactoring PRs to make it possible
to parse CFI programs from raw bytes without build-time dependencies
on lower-level libraries to provide relocations.

Today, CFIProgram and DWARFDataProvider work without relocation
providers by using if-statements. But this approach still requires
the build-time dependency on the relocation provider.

After this series of PRs dwarf::CFIProgram will be usable from the
low-level libraries themselves without creating circular dependencies.
